+++ Sources/KeyboardKit/Actions/KeyboardActionHandler.swift
@@ -19,25 +19,11 @@ import UIKit
 public protocol KeyboardActionHandler: AnyObject {
     
     func handle(_ gesture: KeyboardGesture, on action: KeyboardAction, sender: Any?)
-    
-    @available(*, deprecated, message: "Use handle(_ gesture:on:sender:) instead")
-    func handleTap(on action: KeyboardAction, view: UIView)
-    
-    @available(*, deprecated, message: "Use handle(_ gesture:on:sender:) instead")
-    func handleRepeat(on action: KeyboardAction, view: UIView)
-    
-    @available(*, deprecated, message: "Use handle(_ gesture:on:sender:) instead")
-    func handleLongPress(on action: KeyboardAction, view: UIView)
 }
 
 public extension KeyboardActionHandler {
     
     func handle(_ gesture: KeyboardGesture, on action: KeyboardAction) {
         handle(gesture, on: action, sender: nil)
-    }
-    
-    @available(*, deprecated, message: "Use handle(_ gesture:on:sender:) instead")
-    func handle(_ gesture: KeyboardGesture, on action: KeyboardAction, view: UIView) {
-        handle(gesture, on: action, sender: view)
     }
 }